+++ src/Header/ContentLength.php
@@ -4,24 +4,45 @@ declare(strict_types = 1);
 namespace Innmind\Http\Header;
 
 use Innmind\Http\Header as HeaderInterface;
-use Innmind\Immutable\Sequence;
+use Innmind\Immutable\{
+    Sequence,
+    Maybe,
+};
 
 /**
  * @psalm-immutable
  */
 final class ContentLength implements HeaderInterface
 {
-    public function __construct(
-        private ContentLengthValue $value,
+    /**
+     * @param int<0, max> $length
+     */
+    private function __construct(
+        private int $length,
     ) {
     }
 
     /**
      * @psalm-pure
+     *
+     * @param int<0, max> $length
      */
     public static function of(int $length): self
     {
-        return new self(new ContentLengthValue($length));
+        return new self($length);
+    }
+
+    /**
+     * @psalm-pure
+     *
+     * @return Maybe<self>
+     */
+    public static function maybe(int $length): Maybe
+    {
+        return Maybe::of(match (true) {
+            $length >= 0 => new self($length),
+            default => null,
+        });
     }
 
     #[\Override]
@@ -37,11 +58,11 @@ final class ContentLength implements HeaderInterface
     }
 
     /**
-     * @return 0|positive-int
+     * @return int<0, max>
      */
     public function length(): int
     {
-        return $this->value->length();
+        return $this->length;
     }
 
     #[\Override]
@@ -52,6 +73,6 @@ final class ContentLength implements HeaderInterface
 
     private function header(): Header
     {
-        return new Header('Content-Length', $this->value);
+        return new Header('Content-Length', new Value\Value((string) $this->length));
     }
 }